# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
  description: 'TaskY API — canonical contract generated from springdoc on a dev-profile boot (Observed). Regenerate: boot
    api with dev profile, GET /api-docs, convert to YAML.'
  x-generated: 2026-09-26, feat OIDC code exchange + timer e2e
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/projects/{projectId}:
    get:
      tags:
      - project-controller
      operationId: get_1
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProjectResponse'
    put:
      tags:
      - project-controller
      operationId: update_2
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProjectResponse'
    delete:
      tags:
      - project-controller
      operationId: delete_2
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/organizations/{orgId}/departments/{deptId}:
    put:
      tags:
      - department-controller
      operationId: update_4
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: deptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
    delete:
      tags:
      - department-controller
      operationId: delete_4
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: deptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/projects/{projectId}/cross-department-access:
    get:
      tags:
      - cross-department-access-controller
      operationId: list_3
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CrossDepartmentAccessResponse'
    post:
      tags:
      - cross-department-access-controller
      operationId: grant
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
                format: uuid
        required: true
      responses:
        '200':
          description: OK
...
  /api/v1/projects/{projectId}/assignments:
    get:
      tags:
      - project-assignment-controller
      operationId: list_5
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectAssignmentResponse'
    post:
      tags:
      - project-assignment-controller
      operationId: assign
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
                format: uuid
        required: true
      responses:
        '200':
          description: OK
...
  /api/v1/departments/{deptId}/projects:
    post:
      tags:
      - project-controller
      operationId: create_7
      parameters:
      - name: deptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProjectResponse'
...
  /api/v1/reports/financials/projects:
    get:
      tags:
      - report-controller
      operationId: projectFinancials
      parameters:
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: projectId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: membershipId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectFinancialResponse'
...
  /api/v1/organizations/{orgId}/projects:
    get:
      tags:
      - project-controller
      operationId: listByOrganization
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
...
  /api/v1/projects/{projectId}/cross-department-access/{departmentId}:
    delete:
      tags:
      - cross-department-access-controller
      operationId: remove
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: departmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
  /api/v1/projects/{projectId}/assignments/{membershipId}:
    delete:
      tags:
      - project-assignment-controller
      operationId: remove_1
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
...
components:
  schemas:
...
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        color:
          type: string
        managerMembershipId:
          type: string
          format: uuid
        hourlyRate:
          type: number
        estimatedSeconds:
          type: integer
          format: int64
        budgetSeconds:
          type: integer
          format: int64
        budgetAmount:
          type: number
        isActive:
          type: boolean
        expectedVersion:
          type: integer
          format: int64
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        departmentId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        color:
          type: string
        managerMembershipId:
          type: string
          format: uuid
        hourlyRate:
          type: number
        estimatedSeconds:
          type: integer
          format: int64
        budgetSeconds:
          type: integer
          format: int64
        budgetAmount:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
...
    CreateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
    DepartmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
        managerMembershipId:
          type: string
          format: uuid
        hourlyRate:
          type: number
        estimatedSeconds:
          type: integer
          format: int64
        budgetSeconds:
          type: integer
          format: int64
        budgetAmount:
          type: number
...
    ProjectFinancialResponse:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
        projectName:
          type: string
        estimatedSeconds:
          type: integer
          format: int64
        actualApprovedSeconds:
          type: integer
          format: int64
        actualNotApprovedSeconds:
          type: integer
          format: int64
        remainingSeconds:
          type: integer
          format: int64
        progressPercent:
          type: number
          format: double
        budgetSeconds:
          type: integer
          format: int64
        budgetAmount:
          type: number
        cost:
          type: number
        revenue:
          type: number
        margin:
          type: number
...
    CrossDepartmentAccessResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        departmentId:
          type: string
          format: uuid
        grantedBy:
          type: string
          format: uuid
        grantedAt:
          type: string
          format: date-time
    ProjectAssignmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        membershipId:
          type: string
          format: uuid
        assignedAt:
          type: string
          format: date-time
...